+++ modules/images.py
@@ -553,11 +553,13 @@ def save_image(image, path, basename = '', seed=None, prompt=None, extension=sha
         return None, None
     if not check_grid_size([image]):
         return None, None
-    if path is None or len(path) == 0: # set default path to avoid errors when functions are triggered manually or via api and param is not set
+    if path is None or len(path) == 0:
         path = shared.opts.outdir_save
+
+    # namegen
     namegen = FilenameGenerator(p, seed, prompt, image, grid=grid)
     if shared.opts.save_to_dirs:
-        dirname = namegen.apply(shared.opts.directories_filename_pattern or "[prompt_words]")
+        dirname = namegen.apply(shared.opts.directories_filename_pattern or "[date]")
         path = os.path.join(path, dirname)
     if forced_filename is None:
         if short_filename or seed is None:
@@ -565,21 +567,9 @@ def save_image(image, path, basename = '', seed=None, prompt=None, extension=sha
         if shared.opts.samples_filename_pattern and len(shared.opts.samples_filename_pattern) > 0:
             file_decoration = shared.opts.samples_filename_pattern
         else:
-            file_decoration = "[seq]-[prompt_words]"
+            file_decoration = "[seq]-[model_name]-[prompt_words]"
         file_decoration = namegen.apply(file_decoration)
-        file_decoration += suffix
-        if shared.opts.save_images_add_number:
-            if '[seq]' not in file_decoration:
-                file_decoration = f"[seq]-{file_decoration}"
-            basecount = get_next_sequence_number(path, basename)
-            filename = None
-            for i in range(9999):
-                seq = f"{basecount + i:05}" if basename == '' else f"{basename}-{basecount + i:04}"
-                filename = os.path.join(path, f"{file_decoration.replace('[seq]', seq)}.{extension}")
-                if not os.path.exists(filename):
-                    break
-        else:
-            filename = os.path.join(path, f"{file_decoration}.{extension}") if basename == '' else os.path.join(path, f"{basename}-{file_decoration}.{extension}")
+        filename = os.path.join(path, f"{file_decoration}{suffix}.{extension}") if basename is None or basename == '' else os.path.join(path, f"{basename}-{file_decoration}{suffix}.{extension}")
     else:
         filename = os.path.join(path, f"{forced_filename}.{extension}")
     pnginfo = existing_info or {}
@@ -589,6 +579,21 @@ def save_image(image, path, basename = '', seed=None, prompt=None, extension=sha
     params.filename = namegen.sanitize(filename)
     dirname = os.path.dirname(params.filename)
     os.makedirs(dirname, exist_ok=True)
+
+    # sequence
+    if shared.opts.save_images_add_number or '[seq]' in params.filename:
+        if '[seq]' not in params.filename:
+            params.filename = f"[seq]-{params.filename}"
+        basecount = get_next_sequence_number(dirname, basename)
+        for i in range(9999):
+            seq = f"{basecount + i:05}" if basename == '' else f"{basename}-{basecount + i:04}"
+            filename = params.filename.replace('[seq]', seq)
+            if not os.path.exists(filename):
+                debug(f'Prompt sequence: input="{params.filename}" seq={seq} output="{filename}"')
+                params.filename = filename
+                break
+
+    # callbacks
     script_callbacks.before_image_saved_callback(params)
     exifinfo = params.pnginfo.get('UserComment', '')
     exifinfo = (exifinfo + ', ' if len(exifinfo) > 0 else '') + params.pnginfo.get(pnginfo_section_name, '')